+++ integration_tests/suite/helpers/fixtures.py
@@ -39,7 +39,7 @@ def http_tenant(**tenant_args):
     return decorator
 
 
-def http_user(**user_args):
+def http_user_register(**user_args):
     if 'username' not in user_args:
         user_args['username'] = _random_string(20)
     if 'password' not in user_args:
@@ -50,7 +50,7 @@ def http_user(**user_args):
     def decorator(decorated):
         @wraps(decorated)
         def wrapper(self, *args, **kwargs):
-            user = self.client.users.new(**user_args)
+            user = self.client.users.register(**user_args)
             try:
                 result = decorated(self, user, *args, **kwargs)
             finally: