+++ wazo_auth/services.py
@@ -234,11 +234,13 @@ class UserService(_Service):
         return self._dao.user.list_(**kwargs)
 
     def new_user(self, **kwargs):
-        password = kwargs.pop('password')
-        salt, hash_ = self._encrypter.encrypt_password(password)
+        password = kwargs.pop('password', None)
+        if password:
+            kwargs['salt'], kwargs['hash_'] = self._encrypter.encrypt_password(password)
+
         logger.info('creating a new user with params: %s', kwargs)  # log after poping the password
-        # a confirmation email should be sent
-        return self._dao.user.create(salt=salt, hash_=hash_, **kwargs)
+        # a confirmation email should be sent if the email is not confirmed
+        return self._dao.user.create(**kwargs)
 
     def remove_policy(self, user_uuid, policy_uuid):
         nb_deleted = self._dao.user.remove_policy(user_uuid, policy_uuid)
@@ -255,6 +257,9 @@ class UserService(_Service):
         try:
             hash_, salt = self._dao.user.get_credentials(username)
         except exceptions.UnknownUsernameException:
+            return False
+
+        if not hash_ or not salt:
             return False
 
         return hash_ == self._encrypter.compute_password_hash(password, salt)