+++ integration_tests/suite/test_database.py
@@ -820,13 +820,21 @@ class TestUserDAO(_BaseDAOTestCase):
         hash_ = 'the_hashed_password'
         email_address = 'foobar@example.com'
 
-        user_uuid = self._user_dao.create(username, email_address, hash_, self.salt)['uuid']
+        user_uuid = self._user_dao.create(username, email_address, hash_=hash_, salt=self.salt)['uuid']
         try:
             assert_that(user_uuid, equal_to(ANY_UUID))
             result = self._user_dao.list_(uuid=user_uuid)
             assert_that(result, contains(has_entries(
                 username=username,
                 emails=contains(has_entries(address=email_address, confirmed=False, main=True)))))
+            assert_that(
+                calling(self._user_dao.create).with_args('foo', 'foo@bar.baz', uuid=user_uuid, hash_='', sale=''),
+                raises(
+                    exceptions.ConflictException,
+                    has_properties(
+                        'status_code', 409,
+                        'resource', 'users',
+                        'details', has_entries('uuid', ANY))))
         finally:
             self._user_dao.delete(user_uuid)
 
@@ -835,7 +843,7 @@ class TestUserDAO(_BaseDAOTestCase):
         hash_ = 'the_hashed_password'
         email_address = 'foobar@example.com'
 
-        user_uuid = self._user_dao.create(username, email_address, hash_, self.salt,
+        user_uuid = self._user_dao.create(username, email_address, hash_=hash_, salt=self.salt,
                                           email_confirmed=True)['uuid']
 
         result = self._user_dao.list_(uuid=user_uuid)
@@ -849,7 +857,7 @@ class TestUserDAO(_BaseDAOTestCase):
     @fixtures.user(username='foobar')
     def test_that_the_username_is_unique(self, user_uuid):
         assert_that(
-            calling(self._user_dao.create).with_args('foobar', 'foo@bar', 'hash_two', self.salt),
+            calling(self._user_dao.create).with_args('foobar', 'foo@bar', hash_='hash_two', salt=self.salt),
             raises(
                 exceptions.ConflictException,
                 has_properties(
@@ -863,7 +871,7 @@ class TestUserDAO(_BaseDAOTestCase):
     @fixtures.user(email_address='foobar@example.com')
     def test_that_the_email_is_unique(self, user_uuid):
         assert_that(
-            calling(self._user_dao.create).with_args('bar', 'foobar@example.com', 'hash_two', self.salt),
+            calling(self._user_dao.create).with_args('bar', 'foobar@example.com', hash_='hash_two', salt=self.salt),
             raises(
                 exceptions.ConflictException,
                 has_properties(