+++ wazo_auth/plugins/backends/wazo_user.py
@@ -4,12 +4,12 @@
 
 import logging
 
-from wazo_auth import ACLRenderingBackend, BaseAuthenticationBackend
+from wazo_auth import UserAuthenticationBackend
 
 logger = logging.getLogger(__name__)
 
 
-class WazoUser(BaseAuthenticationBackend, ACLRenderingBackend):
+class WazoUser(UserAuthenticationBackend):
 
     def load(self, dependencies):
         super(WazoUser, self).load(dependencies)
@@ -26,23 +26,21 @@ class WazoUser(BaseAuthenticationBackend, ACLRenderingBackend):
         return self.render_acl(acl_templates, self.get_user_data, username=username)
 
     def get_ids(self, username, args):
-        return self._get_user_uuid(username), None
+        uuid = self._get_user_uuid(username)
+        return uuid, uuid
 
     def verify_password(self, username, password, args):
         return self._user_service.verify_password(username, password)
 
     def get_user_data(self, *args, **kwargs):
         user_uuid = self._get_user_uuid(kwargs['username'])
+        confd_data = super(WazoUser, self).get_user_data(uuid=user_uuid)
         tenants = self._user_service.list_tenants(user_uuid)
         groups = self._user_service.list_groups(user_uuid)
         for group in groups:
             group['users'] = self._group_service.list_users(group['uuid'])
 
-        return {
-            'username': kwargs['username'],
-            'tenants': tenants,
-            'groups': groups,
-        }
+        return dict(username=kwargs['username'], tenants=tenants, groups=groups, **confd_data)
 
     def _get_user_uuid(self, username):
         matching_users = self._user_service.list_users(username=username)