+++ CHANGELOG.md
@@ -55,7 +55,7 @@ The user creation url now needs a token and does not require a password anymore
   * DELETE ``0.1/tenants/<tenant_uuid>/users/<user_uuid>``
   * GET ``0.1/tenants/<tenant_uuid>/users/``
   * GET ``0.1/users/<user_uuid>/tenants/``
-  
+
 * A new ressource has been added to manage Groups
 
   * POST `0.1/groups`